if len(url) != 2, raise error

+++ github_viewer_slack/github_json_maker/rm.py
@@ -7,4 +7,8 @@ import json
 
 @respond_to('^\s*rm\s+(\S+)\s*')
 def rm(message, github_url):
+    if len(github_url.split('/')) != 2:
+        message.reply('Illegal as github repository')
+        return
+    user_name, repo_name = github_url.split('/')
     message.reply("rm message")
